Moved files into folders, moved google API to another file, added some items to render text correctly
