+++ client/src/components/camera/camera.css
