+++ client/src/components/camera/capture.js
