+++ client/src/components/camera/home.js
