+++ client/src/components/login/verify-info.js
